refactoring userSlice, todoSlice and respective hooks and store
