+++ frontend/src/hooks/useToDoById.js
@@ -1,12 +1,12 @@
 import { useDispatch, useSelector } from "react-redux";
-import { deleteToDo, fetchToDoById, updateToDo } from "../redux/userSlice";
+import { deleteToDo, fetchToDoById, updateToDo } from "../redux/todoSlice";
 import { useEffect } from "react";
 
 export const useToDoById = (id) => {
     const dispatch = useDispatch();
-    const data = useSelector((state) => state.user.toDo);
-    const status = useSelector((state) => state.user.toDoStatus);
-    const error = useSelector((state) => state.user.toDoError);
+    const data = useSelector((state) => state.todo.toDo);
+    const status = useSelector((state) => state.todo.toDoStatus);
+    const error = useSelector((state) => state.todo.toDoError);
 
     useEffect(() => {
         if (status === "idle") {
