+++ frontend/src/hooks/useToDos.js
@@ -1,12 +1,12 @@
 import { useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import { addNewToDo, fetchToDos, updateToDo } from "../redux/userSlice.js";
+import { addNewToDo, fetchToDos } from "../redux/todoSlice.js";
 
 export const useToDos = () => {
     const dispatch = useDispatch();
-    const data = useSelector((state) => state.user.toDos);
-    const status = useSelector((state) => state.user.toDosStatus);
-    const error = useSelector((state) => state.user.toDosError);
+    const data = useSelector((state) => state.todo.toDos);
+    const status = useSelector((state) => state.todo.toDosStatus);
+    const error = useSelector((state) => state.todo.toDosError);
 
     useEffect(() => {
         if (status === "idle") {
